Using struct everywhere now and new widget started
Had to explicitly register MP stuff as using the [SyncWorker] attribute
will cause a crash `* Assertion at method-to-ir.c:14035, condition
`load_opcode != OP_LOADV_MEMBASE' not met` on RimWorld game startup!

+++ Source/State.cs
@@ -11,46 +11,60 @@ using Multiplayer.API;
 using Verse;
 
 // Implement a static state storage database using Set/Get/Del/Exists kind of like redis
-// Make sure to SyncMethod for any method that changes internal state for multiplayer
+// Make sure to register SyncMethod for any method that mutates state for multiplayer syncing
 namespace KanbanStockpile
 {
+    public struct KanbanSettings
+    {
+        public int srt; // stack refill threshold [0,100]
+        public int ssl; // similar stack limit [0,15] (higher values in big stockpiles will eat CPU)
+    }
+
     static class State
     {
-        public struct ExtraConfig
-        {
-            public int srt; // stack refill threshold [0,100]
-            public int ssl; // similar stack limit [0,256] (higher values in big stockpiles will eat CPU)
-        }
-
         private static int defaultStackRefillThreshold = 100;
         private static int defaultSimilarStackLimit = 0;
 
         // key is a string from Verse.Zone.label (or pass in "___clipboard" as clipboard's owner is null)
         // val is an int whole number in range [0,100]
-        private static Dictionary<string, ExtraConfig> db = new Dictionary<string, ExtraConfig>();
+        private static Dictionary<string, KanbanSettings> db = new Dictionary<string, KanbanSettings>();
 
         public static bool Exists(string label) {
             return db.ContainsKey(label);
         }
 
-        public static ExtraConfig Get(string label) {
+        public static KanbanSettings Get(string label) {
             if(db.ContainsKey(label)) {
                 return db[label];
             }
-            ExtraConfig result;
+            KanbanSettings result;
             result.srt = defaultStackRefillThreshold;
             result.ssl = defaultSimilarStackLimit;
             return result;
         }
 
-        [SyncMethod]
-        public static void Set(string label, ExtraConfig ec) {
-            db[label] = ec;
+        //[SyncMethod]
+        public static void Set(string label, KanbanSettings ks) {
+            db[label] = ks;
         }
 
-        [SyncMethod]
+        //[SyncMethod]
         public static void Del(string label) {
-            db.Remove(label);
+            if(db.ContainsKey(label)) {
+                db.Remove(label);
+            }
+        }
+
+        //[SyncWorker(shouldConstruct = false)]
+        public static void SyncKanbanSettings(SyncWorker sync, ref KanbanSettings ks)
+        {
+            sync.Bind(ref ks.srt);
+            sync.Bind(ref ks.ssl);
+            //if (sync.isWriting) {
+            //    sync.Write(ks);
+            //} else {
+            //    ks = sync.Read<KanbanSettings>();
+            //}
         }
     }
 }
